Move remaining utils in `breadboard` to `utils`.
