+++ packages/breadboard/src/inspector/graph/node-describer-manager.ts
@@ -22,11 +22,11 @@ import type {
   NodeTypeDescriberOptions,
   NodeTypeIdentifier,
 } from "@breadboard-ai/types";
+import { SchemaDiffer } from "@breadboard-ai/utils";
 import {
   assetsFromGraphDescriptor,
   envFromGraphDescriptor,
 } from "../../file-system/assets.js";
-import { SchemaDiffer } from "../../utils/schema-differ.js";
 import { contextFromMutableGraph } from "../graph-store.js";
 import { UpdateEvent } from "./event.js";
 import { GraphDescriptorHandle } from "./graph-descriptor-handle.js";
