+++ packages/utils/src/index.ts
@@ -13,3 +13,10 @@ export * from "./graph-url-like.js";
 export * from "./run-imperative-graph.js";
 export * from "./chunk-repair.js";
 export * from "./schema.js";
+export * from "./hash.js";
+export * from "./promise-queue.js";
+export * from "./relative-path.js";
+export * from "./schema-differ.js";
+export * from "./semver.js";
+export * from "./snapshot-updater.js";
+export * from "./throttler.js";
